【蓝军渗透测试提单】【msprof】【安全】【一般】connection.cpp中InsertCmd函数中tablename未做空字符串校验
# message auto-generated for no-merge-commit merge:
merge main into master

【蓝军渗透测试提单】【msprof】【安全】【一般】connection.cpp中InsertCmd函数中tablename未做空字符串校验

Created-by: msprofDeveloper

Match-id-8a4aba0ee98e4b88e70660f38fa01f7e92fdceaf

+++ analysis/csrc/viewer/database/db_runner.cpp
@@ -35,6 +35,10 @@ static std::string GetColumnsString(const std::vector <TableColumn> &cols)
 
 bool DBRunner::CreateTable(const std::string &tableName, const std::vector <TableColumn> &cols) const
 {
+    if (tableName.empty()) {
+        ERROR("The tableName is empty string");
+        return false;
+    }
     INFO("Start create %", tableName);
     std::string valuesStr = GetColumnsString(cols);
     std::string sql = "CREATE TABLE IF NOT EXISTS " + tableName + " (" + valuesStr + ");";
@@ -53,6 +57,10 @@ bool DBRunner::CreateTable(const std::string &tableName, const std::vector <Tabl
 
 bool DBRunner::DropTable(const std::string &tableName) const
 {
+    if (tableName.empty()) {
+        ERROR("The tableName is empty string");
+        return false;
+    }
     INFO("Start drop %", tableName);
     std::string sql = "DROP TABLE " + tableName + ";";
     auto conn = std::make_shared<Connection>(path_);
